Upgrade.rs for the profile canister.
